+++ CHANGELOG.rst
@@ -38,6 +38,9 @@ Changelog
 * Containers in home page have direct url without channel, example: site.com/content_slug.html instead of site.com/home/content_slug.html
 * Fix url pattern from flatpages, now accept slugs with dashs
 * Add `get_custom_field_value` template tag
+* Fix None hat field on Mirror creation
+* Fix `main_image` caption population on Albums
+* Add new `hat` field on Channel model
 
 0.2.2
 =====